fix: 新增图表dialog改为组件 # Reviewed, transaction id: 29296
